Add full source code format+lint config
`--config=check-darker.toml` can now be used to run Darker on Graylint
with full reformatters and linters on modifications made to the code.
Used reformatters and linters are included when installing with the
`[test]` extra.

The minimum ruff version is now 0.0.292, and its configuration in
`pyproject.toml` has been updated accordingly.